array iterator

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,7 +1,7 @@
 #include "function_pointers.h"
 
 /**
-*array_iterator- function that executes a function given as a
+*array_iterator - function that executes a function given as a
 *parameter on each element of an array.
 *@array: array
 *@size: size of the array.
@@ -15,7 +15,7 @@ void array_iterator(int *array, size_t size, void (*action)(int))
 
 	if (array && action)
 	{
-		for (i = 0; i < size; i++)
+		for (i = 0; i < (int)size; i++)
 			action(array[i]);
 	}
 }
